requirements.txt
fastapi
uvicorn
motor
pika
pydantic
python-dotenv
